[#11] refactor: extract image saving into src/output-image.ts
